refactor(PN-6): split aiService.ts (29KB) into 7 sub-files + barrel
Split ai-service/ directory:
  as-types.ts       — all type/interface exports + handleRateLimitError
  as-chat.ts        — chat, chatText, explainConcept (RAG chat)
  as-generate.ts    — generateFlashcard, generateQuizQuestion (basic)
  as-generate-smart.ts — generateSmart, preGenerate (Fase 8)
  as-reports.ts     — reportContent, resolveReport, getReportStats, getReports
  as-analytics.ts   — submitRagFeedback, getRagAnalytics, getEmbeddingCoverage
  as-ingest.ts      — ingestPdf, ingestEmbeddings, reChunk

aiService.ts → barrel re-exporter (same pattern as platformApi/studentApi)

Quality fixes included:
  - 5x catch(err:any) → catch(err:unknown) + safe Error casting
  - handleRateLimitError now accepts unknown (not just Error)
  - 2x deprecated console.warn wrapped in import.meta.env.DEV guard
  - listModels deprecated function preserved in as-legacy.ts
